+++ cosmic/tests/test_sample.py
@@ -7,6 +7,7 @@ import os
 import unittest
 import numpy as np
 import pandas as pd
+from ..sample import InitialBinaryTable
 from ..sample.sampler.independent import Sample
 from ..sample.sampler.multidim import MultiDim
 from ..sample.sampler.cmc import CMCSample
@@ -400,6 +401,85 @@ class TestSample(unittest.TestCase):
         # Check that the kstar is selected properly
         kstar = MULTIDIMSAMPLECLASS.set_kstar(pd.DataFrame([1.0, 1.0, 1.0, 1.0, 1.0]))
         self.assertEqual(np.mean(kstar), KSTAR_SOLAR)
+
+    def test_sampling_binfrac_zero(self):
+        # check that you can't sample based on size with a binary fraction of 0.0
+        it_fails = False
+        try:
+            InitialBinaryTable.sampler('independent', np.arange(16), np.arange(16),
+                                       primary_model='kroupa01', ecc_model='thermal',
+                                       porb_model='sana12', binfrac_model=0.0,
+                                       SF_start=10.0, SF_duration=0.0, met=0.02,
+                                       size=1000)
+        except ValueError:
+            it_fails = True
+        self.assertTrue(it_fails)
+
+    def test_sampling_targets_bad_input(self):
+        # check that you get an error if you don't supply a size or total_mass
+        it_fails = False
+        try:
+            InitialBinaryTable.sampler('independent', np.arange(16), np.arange(16),
+                                       primary_model='kroupa01', ecc_model='thermal',
+                                       porb_model='sana12', binfrac_model=0.5,
+                                       SF_start=10.0, SF_duration=0.0, met=0.02,
+                                       sampling_target="total_mass")
+        except ValueError:
+            it_fails = True
+        self.assertTrue(it_fails)
+    
+        it_fails = False
+        try:
+            InitialBinaryTable.sampler('independent', np.arange(16), np.arange(16),
+                                       primary_model='kroupa01', ecc_model='thermal',
+                                       porb_model='sana12', binfrac_model=0.5,
+                                       SF_start=10.0, SF_duration=0.0, met=0.02,
+                                       total_mass=None, sampling_target="total_mass")
+        except ValueError:
+            it_fails = True
+        self.assertTrue(it_fails)
+    
+        it_fails = False
+        try:
+            InitialBinaryTable.sampler('independent', np.arange(16), np.arange(16),
+                                       primary_model='kroupa01', ecc_model='thermal',
+                                       porb_model='sana12', binfrac_model=0.5,
+                                       SF_start=10.0, SF_duration=0.0, met=0.02,
+                                       size=None, total_mass=None, sampling_target="size")
+        except ValueError:
+            it_fails = True
+        self.assertTrue(it_fails)
+
+    def test_sampling_targets_size(self):
+        # check that you can sample based on size
+        for size in np.random.randint(100, 1000, size=100):
+            samples = InitialBinaryTable.sampler('independent', np.arange(16), np.arange(16),
+                                                primary_model='kroupa01', ecc_model='thermal',
+                                                porb_model='sana12', binfrac_model=0.5,
+                                                SF_start=10.0, SF_duration=0.0, met=0.02,
+                                                size=size)
+            self.assertGreaterEqual(len(samples[0]), size)
+
+    def test_sampling_targets_mass(self):
+        # check that you can sample based on total mass
+        for mass in np.random.randint(100, 1000, size=100):
+            samples = InitialBinaryTable.sampler('independent', np.arange(16), np.arange(16),
+                                                primary_model='kroupa01', ecc_model='thermal',
+                                                porb_model='sana12', binfrac_model=0.5,
+                                                SF_start=10.0, SF_duration=0.0, met=0.02,
+                                                size=None, total_mass=mass, sampling_target="total_mass")
+            self.assertGreaterEqual(samples[1] + samples[2], mass)
+
+    def test_sampling_targets_mass_trimmed(self):
+        # check that you can sample based on total mass and that the samples are trimmed properly
+        for mass in np.random.randint(100, 1000, size=100):
+            samples = InitialBinaryTable.sampler('independent', np.arange(16), np.arange(16),
+                                                primary_model='kroupa01', ecc_model='thermal',
+                                                porb_model='sana12', binfrac_model=0.5,
+                                                SF_start=10.0, SF_duration=0.0, met=0.02,
+                                                size=None, total_mass=mass, sampling_target="total_mass",
+                                                trim_extra_samples=True)
+            self.assertLessEqual(abs(samples[1] + samples[2] - mass), 300)
 
 class TestCMCSample(unittest.TestCase):
     def test_plummer_profile(self):